Added graphql mutations and queries
