+++ graphql/types/channels.graphql
@@ -1,0 +1,5 @@
+type Channel {
+  appId: String
+  name: String
+  imageUrl: String
+}
