+++ manifest.json
@@ -8,6 +8,7 @@
   "categories": [],
   "dependencies": {},
   "builders": {
+    "graphql": "1.x",
     "node": "6.x",
     "docs": "0.x",
     "configuration": "0.x"
