+++ node/package.json
@@ -12,8 +12,7 @@
     "@vtex/test-tools": "^1.0.0",
     "tslint": "^5.14.0",
     "tslint-config-vtex": "^2.1.0",
-    "typescript": "3.8.3",
-    "vtex.channels-setup": "https://savio--appliancetheme.myvtex.com/_v/private/typings/linked/v1/vtex.channels-setup@0.0.1+build1591393320/public/_types/react"
+    "typescript": "3.8.3"
   },
   "scripts": {
     "lint": "tsc --noEmit --pretty && tslint -c tslint.json --fix './**/*.ts'"
