+++ node/resolvers/mutation/acceptRequest.ts
@@ -1,0 +1,11 @@
+import { acceptRequestResource } from '../../resources/acceptRequest'
+
+export const acceptRequestResolver = (
+  _: unknown,
+  { sellerAccount }: AcceptParams,
+  ctx: Context
+) => acceptRequestResource(sellerAccount, ctx)
+
+interface AcceptParams {
+  sellerAccount: string
+}
