+++ node/resolvers/query/channels.ts
@@ -1,0 +1,4 @@
+import { fetchChannels } from '../../resources/fetchChannels'
+
+export const channelsResolver = (_: unknown, __: unknown, ctx: Context) =>
+  fetchChannels(ctx)
