+++ node/resolvers/query/requests.ts
@@ -1,0 +1,11 @@
+import { fetchRequests } from '../../resources/fetchRequests'
+
+export const requestsResolver = (
+  _: unknown,
+  { as }: RequestsParams,
+  ctx: Context
+) => fetchRequests(as, ctx)
+
+interface RequestsParams {
+  as: string
+}
